update lookup script
